pyproject.toml
[project]
name = "vibepin"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.2.0",
    "cachetools>=5.3.0",
    "pillow>=10.0.0",
    "ultralytics>=8.1.0",
    "huggingface-hub>=0.20.0",
    "transformers>=4.38.0",
    "numpy>=1.24.0",
]

[project.scripts]
vibepin = "vibepin.main:main"
